+++ Cuda/Src/header/fillEns.h
@@ -14,6 +14,6 @@
 
 __global__ void Generation(curandState_t *state,uint64_t nbr, uint64_t sqrtNBR,uint64_t *rand);
 __global__ void fillEnsR(curandState_t *state,Couple *R,int *size,uint64_t *Div,int sizeDiv,uint64_t *premList,int k,uint64_t *rand,uint64_t nbr,int *matrix);
-__device__ void isInEnsembleG(uint64_t *ens, uint64_t y,int size, int *res);
+__device__ void isInEnsembleG(uint64_t *ens, uint64_t y,int size, uint64_t *res);
 
 #endif /* FILLENS_H_ */
